Update
- Agent able to navigate to correct landmark in `simple_speaker_listener`. Problem was with pi update step which used gumbel_softmax(..., tau=self.tau) which is wrong since self.tau is for soft-update. gumbel-softmax uses tau=1, the default value.

+++ Common.py
@@ -100,4 +100,4 @@ def onehot_from_logits(logits):
     output = np.zeros(logits.shape)
     for idx, entry in enumerate(logits):
         output[idx][np.argmax(entry)] = 1.0
-    return T.tensor(output, dtype=T.float32, device=T.device('cuda' if T.cuda.is_available() else 'cpu'))
+    return T.tensor(output, dtype=T.double, device=T.device('cuda' if T.cuda.is_available() else 'cpu'))
